+++ include/omw/string.h
@@ -1,7 +1,7 @@
 /*
 author          Oliver Blaser
-date            20.08.2022
-copyright       MIT - Copyright (c) 2022 Oliver Blaser
+date            05.03.2023
+copyright       MIT - Copyright (c) 2023 Oliver Blaser
 */
 
 #ifndef IG_OMW_STRING_H
@@ -17,6 +17,41 @@ copyright       MIT - Copyright (c) 2022 Oliver Blaser
 #include "../omw/int.h"
 #include "../omw/vector.h"
 
+#if OMW_CPPSTD >= OMW_CPPSTD_20
+#define OMW_STDSTRING_CONSTEXPR constexpr
+#else
+#define OMW_STDSTRING_CONSTEXPR 
+#endif
+
+/*! \addtogroup grp_stringLib
+* @{
+*/
+
+//! \name UTF-8 Byte Strings
+/// @{
+#define OMW_UTF8CP_00C4 "\xC3\x84"
+#define OMW_UTF8CP_00D6 "\xC3\x96"
+#define OMW_UTF8CP_00DC "\xC3\x9C"
+#define OMW_UTF8CP_00E4 "\xC3\xA4"
+#define OMW_UTF8CP_00F6 "\xC3\xB6"
+#define OMW_UTF8CP_00FC "\xC3\xBC"
+#define OMW_UTF8CP_2580 "\xE2\x96\x80"
+#define OMW_UTF8CP_2584 "\xE2\x96\x84"
+#define OMW_UTF8CP_2588 "\xE2\x96\x88"
+/// @}
+
+//! \name UFT-8 Byte String Aliases
+/// @{
+#define OMW_UTF8CP_Auml OMW_UTF8CP_00C4
+#define OMW_UTF8CP_Ouml OMW_UTF8CP_00D6
+#define OMW_UTF8CP_Uuml OMW_UTF8CP_00DC
+#define OMW_UTF8CP_auml OMW_UTF8CP_00E4
+#define OMW_UTF8CP_ouml OMW_UTF8CP_00F6
+#define OMW_UTF8CP_uuml OMW_UTF8CP_00FC
+/// @}
+
+/*! @} */
+
 namespace omw
 {
     /*! \addtogroup grp_stringLib
@@ -25,17 +60,17 @@ namespace omw
 
     //! \name UTF-8 Byte Strings
     /// @{
-    const char* const UTF8CP_00C4 = "\xC3\x84";
-    const char* const UTF8CP_00D6 = "\xC3\x96";
-    const char* const UTF8CP_00DC = "\xC3\x9C";
-    const char* const UTF8CP_00E4 = "\xC3\xA4";
-    const char* const UTF8CP_00F6 = "\xC3\xB6";
-    const char* const UTF8CP_00FC = "\xC3\xBC";
-    const char* const UTF8CP_2580 = "\xE2\x96\x80";
-    const char* const UTF8CP_2584 = "\xE2\x96\x84";
-    const char* const UTF8CP_2588 = "\xE2\x96\x88";
+    const char* const UTF8CP_00C4 = OMW_UTF8CP_00C4;
+    const char* const UTF8CP_00D6 = OMW_UTF8CP_00D6;
+    const char* const UTF8CP_00DC = OMW_UTF8CP_00DC;
+    const char* const UTF8CP_00E4 = OMW_UTF8CP_00E4;
+    const char* const UTF8CP_00F6 = OMW_UTF8CP_00F6;
+    const char* const UTF8CP_00FC = OMW_UTF8CP_00FC;
+    const char* const UTF8CP_2580 = OMW_UTF8CP_2580;
+    const char* const UTF8CP_2584 = OMW_UTF8CP_2584;
+    const char* const UTF8CP_2588 = OMW_UTF8CP_2588;
     /// @}
-
+    
     //! \name UFT-8 Byte String Aliases
     /// @{
     const char* const UTF8CP_Auml = omw::UTF8CP_00C4;
@@ -66,20 +101,38 @@ namespace omw
     class StringReplacePair
     {
     public:
-        StringReplacePair();
-        StringReplacePair(const std::string& searchElement, const std::string& replaceElement);
-        StringReplacePair(const char searchElement, const std::string& replaceElement);
-        StringReplacePair(const std::string& searchElement, const char replaceElement);
-        StringReplacePair(const char searchElement, const char replaceElement);
+        StringReplacePair() : m_s(), m_r() {}
+        StringReplacePair(const std::string& searchElement, const std::string& replaceElement) : m_s(searchElement), m_r(replaceElement) {}
+        StringReplacePair(const char searchElement, const std::string& replaceElement) : m_s(1, searchElement), m_r(replaceElement) {}
+        StringReplacePair(const std::string& searchElement, const char replaceElement) : m_s(searchElement), m_r(1, replaceElement) {}
+        StringReplacePair(const char searchElement, const char replaceElement) : m_s(1, searchElement), m_r(1, replaceElement) {}
         virtual ~StringReplacePair() {}
 
-        const std::string& search() const;
-        const std::string& replace() const;
+        const std::string& search() const { return m_s; }
+        const std::string& replace() const { return m_r; }
 
     private:
-        std::string searchElem;
-        std::string replaceElem;
+        std::string m_s;
+        std::string m_r;
     };
+
+
+
+    //! \name omw::string Implementation Functions
+    /// @{
+    inline OMW_STDSTRING_CONSTEXPR bool contains(const std::string& str, char ch) { return (str.find(ch) != std::string::npos); }
+    inline OMW_STDSTRING_CONSTEXPR bool contains(const std::string& str, const char* s) { return (str.find(s) != std::string::npos); }
+#if (OMW_CPPSTD < OMW_CPPSTD_17)
+    inline OMW_STDSTRING_CONSTEXPR bool contains(const std::string& str, const std::string& s) { return (str.find(s) != std::string::npos); }
+#else
+    inline OMW_STDSTRING_CONSTEXPR bool contains(const std::string& str, const std::string_view& sv) { return (str.find(sv) != std::string::npos); }
+#endif
+
+    omw::stringVector_t split(const std::string& str, char delimiter, omw::stringVector_t::size_type maxTokenCount = omw::stringVector_npos);
+    omw::stringVector_t splitLen(const std::string& str, std::string::size_type tokenLength, omw::stringVector_t::size_type maxTokenCount = omw::stringVector_npos);
+    /// @}
+
+
 
     class string : public std::string
     {
@@ -154,13 +207,13 @@ namespace omw
 
     //! \name Convert To String
     /// @{
-    omw::string to_string(int32_t value);
-    omw::string to_string(uint32_t value);
-    omw::string to_string(int64_t value);
-    omw::string to_string(uint64_t value);
-    omw::string to_string(float value);
-    omw::string to_string(double value);
-    omw::string to_string(long double value);
+    inline omw::string to_string(int32_t value) { return std::to_string(value); }
+    inline omw::string to_string(uint32_t value) { return std::to_string(value); }
+    inline omw::string to_string(int64_t value) { return std::to_string(value); }
+    inline omw::string to_string(uint64_t value) { return std::to_string(value); }
+    inline omw::string to_string(float value) { return std::to_string(value); }
+    inline omw::string to_string(double value) { return std::to_string(value); }
+    inline omw::string to_string(long double value) { return std::to_string(value); }
 
     omw::string to_string(bool value, bool asText = true);
     omw::string to_string(const omw::int128_t& value);
@@ -179,7 +232,7 @@ namespace omw
     /// @{
     bool stob(const std::string& boolStr);
 
-    size_t stouz(const std::string& str, size_t* pos = nullptr, int base = 10);
+    size_t stoz(const std::string& str, size_t* pos = nullptr, int base = 10);
 
     std::pair<int32_t, int32_t> stoipair(const std::string& str, char delimiter = pairtos_defaultDelimiter);
     //std::pair<uint32_t, uint32_t> stouipair(const std::string& str, char delimiter = pairtos_defaultDelimiter);